Move VerifyPassword to PasswordHandler class
